requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dlib==19.24.8
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.2.1
Werkzeug==3.1.3
zipp==3.21.0
